calRoutines calculates the H index and JSI index according to the pair-wise comparison of CCF

calRoutines calculates the H index and JSI index according to the pair-wise comparison of CCF

Usage

calRoutines(
  maf,
  PrimaryId,
  patient.id = NULL,
  CCF_cutoff = 0.1,
  pairByTumor = TRUE,
  use.tumorSampleLabel = FALSE,
  maf_drivers = NULL,
  subtitle = "both",
  ...
)

Arguments

maf

Maf or MafList object generated by MesKit::readMaf() function.

PrimaryId

Primary tumor IDs to indicate the primary-metastases relationships.

patient.id

Select the specific patients. Default NULL, all patients are included.

CCF_cutoff

The minimal cutoffs for the present status. The mutations with CCF smaller than CCF_cutoff were consider as absent statuses.

pairByTumor

Compare JSI between different tumors. Default TRUE.

use.tumorSampleLabel

Logical (Default: FALSE). Rename the 'Tumor_Sample_Barcode' by 'Tumor_Sample_Label'.

maf_drivers

Driver information. Two columns are required, including "Mut_ID" and "is.driver". If provided, add the driver mutations in the plot. "Mut_ID" = str_c(Chromosome, Start_Position, Reference_Allele , Tumor_Seq_Allele2, sep = ":")

subtitle

the information shows in the subtitle. Options including "JSI", "Hindex", "both" and "none". Default is "both"

...

Other options passed to MesKit::subMaf().

References

H index. Hu, Z., et al., Quantitative evidence for early metastatic seeding in colorectal cancer. Nat Genet, 2019. 51(7): p. 1113-1122.

JSI index. Hu, Z., et al., Multi-cancer analysis of clonality and the timing of systemic spread in paired primary tumors and metastases. Nat Genet, 2020. 52(7): p. 701-708.

Examples

library(MesKit)
data.type <- "split1"

maf <- readMaf(
  mafFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.%s.mutation.txt", data.type)),
  ccfFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.%s.CCF.txt", data.type)),
  clinicalFile = system.file(package = "MPTevol", "extdata", sprintf("meskit.%s.clinical.txt", data.type)),
  refBuild = "hg19",
  ccf.conf.level = 0.95
)
cal <- calRoutines(
  maf = maf,
  patient.id = "Met1",
  PrimaryId = "Coad",
  pairByTumor = TRUE,
  use.tumorSampleLabel = TRUE,
  subtitle = "both"
)

plot_grid(plotlist = cal$Met1$plist, nrow = 1)
## add driver information
maf_driver <- data.frame(
  Mut_ID = c("5:112170777:CAGA:-", "1:147092680:-:C"),
  is.driver = c(TRUE, TRUE)
)
cal <- calRoutines(
  maf = maf,
  patient.id = "Met1",
  PrimaryId = "Coad",
  pairByTumor = TRUE,
  use.tumorSampleLabel = TRUE,
  subtitle = "both",
  maf_drivers = maf_driver
)

plot_grid(plotlist = cal$Met1$plist, nrow = 1)