inferClonalTrees

This function have two main modules, including inferring the clonal trees and plot the clonal models.

Usage

inferClonalTrees(
  project.names,
  variants,
  vaf.col.names = NULL,
  ccf.col.names = NULL,
  sample.groups = NULL,
  founding.cluster = 1,
  ignore.clusters = NULL,
  cluster.col.name = "cluster",
  clone.colors = NULL,
  subclonal.test.model = "non-parametric",
  cancer.initiation.model = "monoclonal",
  sum.p = 0.05,
  alpha = 0.05,
  weighted = FALSE,
  consensus.tree = TRUE,
  plot.models = TRUE,
  plot.pairwise.CCF = F,
  highlight.note.col.name = NULL,
  highlight = "is.driver",
  highlight.CCF = FALSE
)

Arguments

project.names

the project names used in the output.

variants

data frame of the variants. At least cluster column and VAF or CCF columns are required. Cluster column should contain cluster identities as continuous integer values starting from 1.

vaf.col.names

the column names of samples containing VAF.

ccf.col.names

the column names of samples containing CCF. Note: either setting vaf.col.names or ccf.col.names.

sample.groups

indicate the samples groups. An example is setNames(c("Primary","Primary","Met","Met"), nm = c("P1","P2","M1","M1") )

founding.cluster

the name of founding clones, one of the most important parameters. For most of circumstances, the founding cluster is the cluster with the highest average CCF cluster.

ignore.clusters

the clusters that ignores to analysis. For some clusters, especially clusters that have low vafs in all samples, were probably false-positive clusters.

cluster.col.name

the column names that containing cluster information.

clone.colors

setting clone colors.

subclonal.test.model

What model to use when generating the bootstrap values are: c('non-parametric', 'normal', 'normal-truncated', 'beta', 'beta-binomial'). (Default = "non-parametric")

cancer.initiation.model

cancer evolution model to use, c('monoclonal', 'polyclonal'). Monoclonal model assumes the orginal tumor (eg. primary tumor) arises from a single normal cell; polyclonal model assumes the original tumor can arise from multiple cells (ie. multiple founding clones). In the polyclonal model, the total VAF of the separate founding clones must not exceed 0.5.

sum.p

min probability that the cluster is non-negative in a sample(Default = 0.05).

alpha

alpha level in confidence interval estimate for the cluster (Default = 0.05).

weighted

weighted model (default = FALSE)

consensus.tree

whether build the consensus tree (Default = TRUE).

plot.models

whether plot the models (Default = TRUE).

plot.pairwise.CCF

whether plot pairwise CCF comparison (Default = FALSE).

highlight.note.col.name

highlight context.

highlight

column name to indicate whether highlight the sites (TRUE or FALSE).

highlight.CCF

highlight is CCF or VAF (Default = FALSE).

Details

Inferring the clonal trees is the central process in clonal construction. However, users always find that it is difficult to build clonal trees. Therefore, we should check the cluster structures before building clonal trees. Here are some suggestions about building clonal trees.

chose the optimal clustering methods. Before do mutation clustering. We should removing the low-quality mutations. The indels are suggested to be removed. The mutations in the LOH regions are suggested to be removed. The mutations in the cnv-regions are should be carefully checked.

chose the right founding cluster.

ignore some false-negative clusters. For some clusters, especially clusters that have low vafs in all samples, were probably false-positive clusters. Removing clusters that having too few mutations.

try different cutoffs. The two parameters sum.p and alpha are used to determine whether a cluster is in a sample. A relaxed cutoffs (small values of the two parameters) enables more clusters are though to be present in the sample.